Huge restructure
Sorry i am doing this in one commit!
Ok so I have done a massive restructure of the code... It took a while. Nothing has changed only how the code is stored and how different things are called. The restructure allows for the program to work fine without tkinter without a GUI.
The restructure also allows possible easier expansion of the code. The restructure splits the code into 4 python files:
pico_project.py - This is the start file (Like normal) this is where the program enters and receives any parameters.
pico_project_generation.py - This file was needed when ripped the GUI from the main file to prevent import loops and incomplete import errors.
pico_project_settings.py - in here are all the variables that are needed by all of the other files in the project. I also put things like the code fragments and lists to make them easy to find when adding more. Again, this was needed so the main 2 files pico_project and pico_project_gui can access all the variables needed.
pico_project_gui.py - This contains ALL code needed for the gui. If the user wanted to they could drop this file entirely and the project would/should work fine without a GUI.
If you decide to merge this code, I would highly recommend testing my code and double checking it before merging it with the master.
The code also is not the neatest, but I have tried my best!
I really hope this helps at least a little!
